requirements.txt
pyzabbix>=0.7.4
mysql-connector-python>=8.0.11

